Basic restructure
Did some restructuring and added basic functionality for having some variables be optional in the. main ibpm module

Summary of the re-structure:

ib_state (moved circ2_st_vflux to fluid-operators/lin.jl, State, IBState and init_state to the pre-processing directory, and base_flux! to the timestepping directory)
ib_domain —> fluid-domain/fluid-domain.jl and structure-domain/structure-domain.jl 
The make_naca and make_cylinder functions went into utilities within structure-domain/
ib_problems
Models
multigrid_utils —> fluid-operators/multigrid-utils.jl
ib_matutils —> fluid-operators/indexing-utils.jl
dst_inversion —> fluid-operators/utilities/dst-inversion.jl
Nonlin —> fluid-operators/nonlin.jl
ib_mats —> fluid-operators/lin.jl (also moved the Q, W functions into fluid-operators/nonlin.jl)
ib_problems, ib_state, and models primarily went into the pre-processing directory, in one of either pre-processing.jl, problem-types.jl, 
	Caveats: 
	- Motion_type function (ib_problems.jl) moved to structure-domain/structure-domain.jl
ib_coupling.jl —> interface-couling/interface-coupling.jl
plot_utils.jl —> plotting/plotting-utils.jl
ibpm_solver.jl and Schemes —> timestepping/timestepping.jl (except for the explicit_scheme type, which went into pre-processing/pre-processing.jl)


Some more cleanup todos/questions/thoughts:
Better plotting utilities
Functionality for providing bodies as functions or a set of points that get re-interpolated to have the right grid spacing
Right now IBModel and IBProblem seem kind of arbitrarily defined to me. Is there a more meaningful designation/set of designations?
Make ET more efficient
Condense get_B into a single case that works for both UniformGrid and MultiGrid
Optimize Ainv, RCinv
Eventually, get_AB and get_B should be changed because we won’t want to store B for the FSI case
Combine the multigrid and uniformgrid versions of KKT
Set up KKT as a linear map?
Does get_nonlin() (Line 4, nonlin.jl) get used anywhere??
Where do the various working memory arrays get called? Would be helpful to rename
Do we need to import Base.Threads.@threads???
Do we need to build dst_plans (line 130, pre-processing.jl) after creating the operators?
What is the utility of having Model and Problem be separate?
Meaningful renaming of A and B?
