+++ src/fluid-domain/fluid-domain-include.jl
@@ -1,0 +1,5 @@
+"""
+Information about which files to include within this directory
+"""
+
+include("fluid-domain.jl")
