+++ src/fluid-operators/fluid-operators-include.jl
@@ -1,0 +1,9 @@
+"""
+Information about which files to include within this directory
+"""
+
+include("lin.jl")
+include("nonlin.jl")
+include("utilities/dst-inversion.jl")
+include("utilities/indexing-utils.jl")
+include("utilities/multigrid-utils.jl")
