+++ src/structure-domain/structure-domain-include.jl
@@ -1,0 +1,6 @@
+"""
+Information about which files to include within this directory
+"""
+
+include("structure-domain.jl")
+include("sample-bodies.jl")
